+++ packages/gui/src/components/html/types.ts
@@ -1,7 +1,7 @@
 export interface ElementData {
   type: 'element' | 'text'
   tagName?: string
-  attributes?: Record<string, string>
+  attributes?: Record<string, string | Slot>
   // `style` is an attribute, but we treat it specially for CSS.gui
   style?: Record<string, any>
   value?: string
@@ -17,7 +17,7 @@ export interface Slot {
   name: string
   value?: string
   tagName?: string
-  attributes?: Record<string, string>
+  attributes?: Record<string, string | Slot>
   style?: Record<string, any>
   children?: HtmlNode[]
   props?: Props
@@ -28,7 +28,7 @@ export interface ComponentData {
   tagName: string
   props?: Props
   value: HtmlNode
-  attributes?: Record<string, string>
+  attributes?: Record<string, string | Slot>
   style?: Record<string, any>
   children?: HtmlNode[]
 }