+++ packages/gui/src/lib/codegen/to-react-props.ts
@@ -1,9 +1,10 @@
 import * as propInfo from 'property-information'
+import { stringifySlotInProp } from './util'
 
 const SCHEMA = 'html' as unknown as propInfo.Schema
 
 type Props = Record<string, any>
-export const toReactProps = (props: Props): Props => {
+export const toReactProps = ({ outerProps, ...props }: Props): Props => {
   return Object.entries(props).reduce((acc, curr) => {
     const [key, value] = curr
 
@@ -15,12 +16,12 @@ export const toReactProps = (props: Props): Props => {
       const propName = info.property || key
 
       return {
-        [propName]: value,
+        [propName]: stringifySlotInProp(value, outerProps),
         ...acc,
       }
     } catch (e) {
       return {
-        [key]: value,
+        [key]: stringifySlotInProp(value, outerProps),
         ...acc,
       }
     }