+++ packages/gui/src/lib/codegen/util.ts
@@ -4,7 +4,7 @@ import { visit } from 'unist-util-visit'
 import { HtmlNode, Slot } from '../../components/html/types'
 
 export const getSlots = (value: HtmlNode) => {
-  if (value.type === 'text' || value.type === 'slot') {
+  if (isText(value) || isSlot(value)) {
     return []
   }
 
@@ -13,6 +13,14 @@ export const getSlots = (value: HtmlNode) => {
     .use(() => (tree: any) => {
       visit(tree, 'slot', (node: any) => {
         slots.push(node)
+      })
+      visit(tree, 'element', (node: any) => {
+        const attributes = node.attributes || {}
+        Object.values(attributes).forEach((val: any) => {
+          if (isSlot(val)) {
+            slots.push(val)
+          }
+        })
       })
     })
     .runSync(value)
@@ -30,3 +38,19 @@ export const hasChildrenSlot = (value: HtmlNode) => {
   const slots = getSlots(value)
   return !!slots.find((slot) => slot.name === 'children')
 }
+
+export const stringifySlotInProp = (value: any, outerProps: any) => {
+  if (isSlot(value)) {
+    const slotName = value.name
+    if (outerProps && outerProps[slotName]) {
+      return outerProps[slotName]
+    }
+
+    return value.value ?? null
+  }
+
+  return value
+}
+
+export const isText = (value: HtmlNode) => value?.type === 'text'
+export const isSlot = (value: HtmlNode) => value?.type === 'slot'